[+-] Split up files
I split up the files so they are easier to navigate, and find what you
are looking for based on the file name.
